% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_components.R
\name{adjust_components}
\alias{adjust_components}
\title{this is a fiendish internal function that I wish I had described in greater detail}
\usage{
adjust_components(g, max_iter, force)
}
\description{
this is a fiendish internal function that I wish I had described in greater detail
}
