% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_from_solved_heights.R
\name{angle_from_solved_heights}
\alias{angle_from_solved_heights}
\title{Angle from solved heights}
\usage{
angle_from_solved_heights(height_solution, g = IEEE_118)
}
\arguments{
\item{height_solution}{the height embeddings of the graph}

\item{g}{A igraph object}
}
\description{
This function is borderline for deletion
It calculates the angle between two edges
}
