% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_SETS2.R
\name{auto_SETS2}
\alias{auto_SETS2}
\title{Function that finds the convergence parameters of a simulation}
\usage{
auto_SETS2(g, force = "net_generation", flow = "power_flow",
  distance = "distance", edge_name = "edge_name", tstep = 0.02,
  mass = 1, max_iter = 20000, tol = 0.002, sparse = FALSE,
  two_node_solution = TRUE, restarts = 100, step_size = 0.1,
  sample = 1)
}
\description{
Function that finds the convergence parameters of a simulation
}
