% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Damping_matrix.R
\name{Calc_Damping_matrix}
\alias{Calc_Damping_matrix}
\title{Calculate damping matrix
This function calculates the damping experienced by each node based on it's contact edges and current speed.
This is a helper function and is seldom called on it's own}
\usage{
Calc_Damping_matrix(EdgeNode, v_vect, kvect, mvect)
}
\arguments{
\item{EdgeNode}{The edgenode matrix}

\item{v_vect}{the velocity of each node}

\item{kvect}{the vector of spring stiffness for each edge}

\item{dvect, }{the vector of horizontal distance between the springs.}
}
\description{
Calculate damping matrix
This function calculates the damping experienced by each node based on it's contact edges and current speed.
This is a helper function and is seldom called on it's own
}
