% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_line_strain.R
\name{Calc_line_strain}
\alias{Calc_line_strain}
\title{Calculate line strain}
\usage{
Calc_line_strain(edge_mat, solved_height_mat, merge_order_mat)
}
\arguments{
\item{edge_mat}{Numeric matrix.}

\item{solved_height_mat}{A numeric matrix. This matrix is the result of the Calc_system_dynamics functions}

\item{merge_order_mat}{A numeric matrix. This provides the order that the vectors will be provided in}
}
\description{
This function calculates the line strain characteristics for a graph. It ia a helper function of 
find_stabil_system and is rarely called by the ser. It is a much faster version of'calc_tension_strain'.
}
