% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_logistic_Spring_Area.R
\name{Calc_logistic_Spring_Area}
\alias{Calc_logistic_Spring_Area}
\title{Logistic sprin area}
\usage{
Calc_logistic_Spring_Area(g, Force, Capacity, eta, gamma, kappa)
}
\arguments{
\item{g}{an igraph object}

\item{Force}{a character string that is the name of the edge atribute that is used as force}

\item{Capacity}{a character string that is the ame of the edge atribute for edge capacity}

\item{eta}{the range of possible logistic values}

\item{gamma}{the minimum value the logistic function will produce}

\item{kappa}{the squashing parameter of the logistic function}
}
\description{
Calculate the spring area of an edge based on a logistic curve
}
\details{
This function allows for variable cross sectional area of an edge. This might be because
there is a complex relationship between the flow and it's physical interpretation.
The output of this function is a graph with an area attribute "Area".
}
