% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_spring_area.R
\name{calc_spring_area}
\alias{calc_spring_area}
\title{Calculate the youngs modulus of the spring}
\usage{
calc_spring_area(g, value, minimum_value, range)
}
\arguments{
\item{g}{an igraph object. The graph representing the network}

\item{value}{a character string. The name of the edge atribute that is used as value}

\item{minimum_value}{a numeric value. Indicating the most stretchy value of youngs modulus}

\item{range}{a numeric value. This gives the range of A values above the minimum, the maximum value value.}
}
\description{
This function adds the graph characteristic A which is the cross sectional area of the springy.
}
