% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_spring_constant.R
\name{calc_spring_constant}
\alias{calc_spring_constant}
\title{Calculate the spring constant}
\usage{
calc_spring_constant(g, E = "E", A = "A", distance = "distance")
}
\arguments{
\item{g}{an igraph object. The graph representing the network}

\item{E}{a character string. The youngs modulus of the edge. The default is E}

\item{A}{a character string. The cross sectional area of the line. The defualt is A. see details on values of A}

\item{distance}{A character string. See details on values of distance}
}
\value{
and edge atribute called k with value \code{EA/distance}
}
\description{
This function adds the graph characteristic k which is the spring constant. When A and distance are both set to 1
\code{k=E} and the spring constant is equivalent to youngs modulus.
}
\details{
The values A and distance are edge atributes referring to the crossectional area of the edge and the horizontal distance of the edge,
in other words the distance between the two nodes at each end of the edge. These values can be set to anything the user wishes, they may be
constant or not. However, consider carefully setting the values to anything else other than 1. There needs to be a clear resoning
or the results will be meaningless. 

For example setting the distance of an edge that represents an electrical cable to the distance
of the electrical cable will return very different results when compared to a constant of one. However, the physical distance between two points
does not necessarily have an impact on the loading of the line and so the results would not be interpretable. In contrast setting the distance
metric to be some function of the line resistance may have meaning and be appropriate. As a general rule distance and area should be set to 1.
}
\seealso{
[calc_spring_youngs_modulus]
}
