% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_spring_youngs_modulus.R
\name{calc_spring_youngs_modulus}
\alias{calc_spring_youngs_modulus}
\title{Calculate the youngs modulus of the spring}
\usage{
calc_spring_youngs_modulus(g, flow, capacity, minimum_value, stretch_range)
}
\arguments{
\item{g}{an igraph object. The graph representing the network}

\item{flow}{a character string. The name of the edge atribute that is used as flow}

\item{capacity}{a character string. The ame of the edge atribute for edge capacity}

\item{minimum_value}{a numeric value. Indicating the most stretchy value of youngs modulus}

\item{stretch_range}{a numeric value. This gives the range of k values above the minimum and the point when loading is 100% of capacity}
}
\description{
This function adds the graph characteristic E which is the youngs modulus of each edge in the graph.
}
