% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_System_Dynamics.R
\name{Calc_System_Dynamics}
\alias{Calc_System_Dynamics}
\title{Calculate system dynamics}
\usage{
Calc_System_Dynamics(NodeStatus, EdgeNode, kvect, dvect, tstep = 1,
  frctmultiplier = 1)
}
\arguments{
\item{NodeStatus}{A data frame The current dynamics and forces experienced by the node a data frame.}

\item{EdgeNode}{}

\item{kvect}{A numeric vector of the spring stiffnesses}

\item{dvect}{A numeric vector of the initial distances between the nodes}

\item{tstep}{A numeric value. The time step, measured in seconds, that will be used to calculate the new dynamic state}

\item{frctmultiplier}{A numeric value. Used to set a multiplier on the friction value. Generally leave this alone.}
}
\description{
This is a helper function that calculates the dynamics at time t of all the nodes in the system.
It is seldom called on it's own but is called by other functions.
}
\details{
The output of the function is a dataframe that is the next iteration of the NodeStatus dataframe the function recieves.
}
