% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_System_Dynamics_sparse.R
\name{Calc_System_Dynamics_sparse}
\alias{Calc_System_Dynamics_sparse}
\title{Calculate system dynamics sparse}
\usage{
Calc_System_Dynamics_sparse(NodeStatus, ten_mat, damp_mat, kvect, dvect,
  mat_size, tstep = 1, non_empty_matrix, frctmultiplier = 1)
}
\arguments{
\item{NodeStatus}{A numeric matrix frame The current dynamics and forces experienced by the nodes. The node ID is not included
This means the order of the matrix has to be fixed.}

\item{ten_mat}{A numeric matrix. This matrix must be a zero matrix or an adjacency matrix of dimensions nxn where n is the number of nodes}

\item{damp_mat}{Identical to the ten_mat. This is supposed to help reduce memory, But I don't think it does}

\item{kvect}{A numeric vector of the spring stiffnesses}

\item{dvect}{A numeric vector of the initial distances between the nodes}

\item{mat_size}{a numeric scaler. The dimensions of the square matrix}

\item{tstep}{A numeric value. The time step, measured in seconds, that will be used to calculate the new dynamic state}

\item{non_empty_matrix}{A numeric matrix. This contains the node indexes in the adjacency matrix it has 4 columns
the row, the column, the absolute index the transpose of the absolute index.}

\item{frctmultiplier}{A numeric value. Used to set a multiplier on the friction value. Generally leave this alone.}
}
\description{
This is a helper function that calculates the dynamics at time t of all the nodes in the system.
It is seldom called on it's own but is called by other functions. This function comes in a version for sparse and dense matrices.
}
\details{
The output of the function is a matrix that is the next iteration of the NodeStatus dataframe the function recieves.
}
