% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_SETSe.R
\name{categorical_SETSe}
\alias{categorical_SETSe}
\title{Embedd categorical values}
\usage{
categorical_SETSe(g, force_var, node_names, k = "k", binary = FALSE)
}
\arguments{
\item{k}{}
}
\description{
Performs SETSe on every level of a categorical variable. This function is experimental
}
