% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_balanced_blocks.R
\name{create_balanced_blocks}
\alias{create_balanced_blocks}
\title{Create balanced blocks}
\usage{
create_balanced_blocks(g, force = "net_generation",
  flow = "power_flow")
}
\arguments{
\item{g}{An igraph object. The network for which embeddings will be found}

\item{force}{A character vector. The name of the node attribute that is the force exerted by the nodes}

\item{flow}{A character vector. The name of the edge attribute that contains the flow over each edge}
}
\value{
A list containing all the bi connected component where each component is balanced to have a net force of 0.
}
\description{
Separates the network into a series of bi-connected components that can be solved separately. 
Solving smaller subgraphs using the bi-connected component method reduces the risk of network divergence.
}
\details{
When networks are separated into the bi-connected subgraphs or blocks. The overall network balance needs to be maintained. 
\code{create_balanced_blocks} maintains the balance by summing the net force across the all the nodes that are being removed from
the subgraph. Therefore a node that is an articulation point has a force value equal to the total of all the nodes on the adjacent
bi-connected component.
}
