% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_stabilised_blocks_expanded.R
\name{Create_stabilised_blocks_expanded}
\alias{Create_stabilised_blocks_expanded}
\title{Create stabilosed blocks}
\usage{
Create_stabilised_blocks_expanded(
  g,
  OriginBlock,
  OriginBlock_number,
  force = "force",
  edge_name = "edge_name",
  tstep = 0.1,
  coef_drag = coef_drag,
  tol = 0,
  distance,
  max_iter,
  mass,
  verbose = FALSE,
  sparse
)
}
\arguments{
\item{g}{An igraph object}

\item{OriginBlock}{An Igraph object. This is created by Create_balanced_blocks function}

\item{OriginBlock_number}{An integer. this is the origin block chosen from the
create_stable_blocks function. Usually this will be the largest block.}

\item{force}{A character string. This is the node attribute that contains the force the nodes exert on the network.}

\item{edge_name}{A character string. This is the edge attribute that contains the edge_name of the edges.}

\item{tstep}{A numeric. The time interval used to iterate through the network dynamics.}

\item{coef_drag}{A numeric. This sets the multiplier of friction. Only use if you want to be annoyed and confused}

\item{tol}{A numeric. The tolerance factor for early stopping.}

\item{distance}{A character string. The edge attribute that contains the original/horizontal distance between nodes.}

\item{max_iter}{An integer. The maximum number of iterations before stopping. Larger networks usually need more iterations.}

\item{mass}{A numeric. This is the mass constant of the nodes in normalised networks this is set to 1.}

\item{verbose}{Logical. This value sets whether messages generated during the process are supressed or not.}

\item{sparse}{Logical. Whether or not the function should be run using sparse matrices. must match the actual matrix, 
this could prob be automated}
}
\value{
A dataframe with the height embeddings of the network
}
\description{
decomposes the network into bi-connected components using articulation points. This speeds up the convergence process
and reduces the chances of the SETS algorithm diverging
}
\seealso{
\code{\link{Create_stabilised_blocks}}
}
