% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Tension_matrix.R
\name{Create_Tension_matrix}
\alias{Create_Tension_matrix}
\title{Create tension matrix}
\usage{
Create_Tension_matrix(EdgeNode, zvect, kvect, dvect)
}
\arguments{
\item{zvect}{A numeric vector. The node elevation.}

\item{kvect}{A numeric vector. The vector of spring coefficients between connected nodes.}

\item{dvect}{A numeric vector. The original distance between adjacent nodes.}

\item{Edgenode}{A dataframe. represents the edge node matrix}
}
\description{
Calculates the edge tension in the network. This is a helper function and is seldom called directly
}
