% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Voronoi.R
\name{Create_Voronoi}
\alias{Create_Voronoi}
\title{Create Voronoi}
\usage{
Create_Voronoi(node_values, sfc_map)
}
\arguments{
\item{node_values}{A dataframe containing a Node column a Longitude column a Latitude column as well as any desired attribute columns}

\item{sfc_map}{An sfc object defining the geometry of the area that the nodes are in.}
}
\description{
Creates a voronoi tessalation from a collection of points whilst preserving desired node properties
}
\details{
This possibly outputs a voronoi tesselation, it definately outputs and sf object
}
