% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Voronoi_edge_df.R
\name{Create_Voronoi_edge_df}
\alias{Create_Voronoi_edge_df}
\title{Create a network from a voronoi tessalation}
\usage{
Create_Voronoi_edge_df(Voronoi_sf, Node_name)
}
\arguments{
\item{Voronoi_sf}{An Sf object that describes the voronoi geometry and has a column that give the node that the cell is based around}

\item{Node_name}{Node_name a bare unquoted name of the Node column in the sf object}
}
\description{
This function allows voronoi tessalations to be converted into networks. This is no fun to do hence this function.
}
\details{
This function is useful when you want to turn a collection of spatial points into a network.
First the points are converted into a voronoi tesselation then this function converts them into a network.
}
