% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance.R
\name{Distance}
\alias{Distance}
\title{Distance
Calculates the distance a node will be from the origin for a given time-step}
\usage{
Distance(z, v, a, t0, t1)
}
\arguments{
\item{z}{A numeric value. The distance at current time}

\item{v}{A numeric value. Velcoity at current time}

\item{a}{A numeric value. Acelleration at current time}

\item{t0}{A numeric value. Current time in seconds from simulation start}

\item{t1}{A numeric value. Time at next time step aka t1 = t0 tdelta t}
}
\description{
Distance
Calculates the distance a node will be from the origin for a given time-step
}
