% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_network_balance.R
\name{Find_network_balance}
\alias{Find_network_balance}
\title{Find the network balance}
\usage{
Find_network_balance(g, force = "net_generation", flow = "power_flow",
  distance = "distance", edge_name = "edge_name", tstep = 0.02,
  mass = 1, max_iter = 20000, coef_drag = 1, tol = 1e-06,
  sparse = FALSE, verbose = TRUE, two_node_solution = TRUE)
}
\arguments{
\item{g}{An igraph object}

\item{force}{A character string. This is the node attribute that contains the force the nodes exert on the network.}

\item{flow}{A character string. This is the edge attribute that is the power flow on the edges.}

\item{distance}{A character string. The edge attribute that contains the original/horizontal distance between nodes.}

\item{edge_name}{A character string. This is the edge attribute that contains the edge_name of the edges.}

\item{tstep}{A numeric. The time interval used to iterate through the network dynamics.}

\item{mass}{A numeric. This is the mass constant of the nodes in normalised networks this is set to 1.}

\item{max_iter}{An integer. The maximum number of iterations before stopping. Larger networks usually need more iterations.}

\item{coef_drag}{A numeric. This sets the multiplier of friction. Only use if you want to be annoyed and confused}

\item{tol}{A numeric. The tolerance factor for early stopping.}

\item{sparse}{Logical. Whether or not the function should be run using sparse matrices. must match the actual matrix, this could prob be automated}

\item{verbose}{Logical. This value sets whether messages generated during the process are supressed or not.}

\item{two_node_solution}{Logical. The}
}
\value{
A list of two elements. A data frame with the height embeddings of the network as well as the convergence dynamics dataframe for the network
}
\description{
This simulates the dynamics of the network for a set number of iterations or until convergence which ever is the sooner.
}
\details{
This function is often used in conjunction with \code{Create_stabilised_blocks} and \code{create_balanced_blocks}
}
\seealso{
\code{\link{Create_stabilised_blocks}} \code{\link{create_balanced_blocks}}
}
