% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_network_balance_expanded.R
\name{Find_network_balance_expanded}
\alias{Find_network_balance_expanded}
\title{Find stabil system expanded
This is a special case function which keeps the history of the network dynamics. It is useful for demonstrations. or Parametrizing difficult networks}
\usage{
Find_network_balance_expanded(g, force = "net_generation",
  flow = "power_flow", tstep = 0.5, distance = "Imp", mass = 1,
  maxIter = 20000, frctmultiplier = 1, tol = 1e-10, verbose = TRUE,
  TwoNodeSolution = TRUE)
}
\arguments{
\item{g}{An igraph object. The network}

\item{force}{A character string}

\item{flow}{A character string. the name of the graph attribute that contains the flow information}

\item{tstep}{A numeric. The time in seconds that elapses between each iteration}

\item{distance}{A character string. The name of the graph attribute that contains the graph distance}

\item{mass}{A numeric. The mass in kg of the nodes, this is arbitrary and commonly 1 is used.}

\item{maxIter}{An interger. The maximum nuber of iterations before terminating the simulation}

\item{frctmultiplier}{A numeric. A multplier used to tune the damping. Generally no need to twiddle}

\item{tol}{A numeric. Early termination. If the dynamics of the nodes fall below this value the algortihm will be classed as 
"converged" and the simulation terminates.}

\item{verbose}{Logical value. Whether the function should output messages or run quietly.}

\item{TwoNodeSolution}{A logical Value. Are two node networks solved using Newton-Raphson}
}
\description{
Find stabil system expanded
This is a special case function which keeps the history of the network dynamics. It is useful for demonstrations. or Parametrizing difficult networks
}
