% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindStabilSystem.R
\name{FindStabilSystem}
\alias{FindStabilSystem}
\title{Find stabil system}
\usage{
FindStabilSystem(g, distance, NodeStatus, EdgeNode, flow, kvect, dvect,
  capacity, edge_name = edge_name, tstep, maxIter = 1000,
  frctmultiplier = 1, tol = 1e-10, verbose = TRUE,
  friction_Stop = FALSE)
}
\arguments{
\item{g}{An igraph object. The network}

\item{NodeStatus}{A data frame The current dynamics and forces experienced by the node a data frame.}

\item{EdgeNode}{}

\item{flow}{A character string. This is the edge attribute that is the power flow on the edges.}

\item{kvect}{A numeric vector of the spring stiffnesses}

\item{dvect}{A numeric vector of the initial distances between the nodes}

\item{capacity}{A character string. This is the edge attribute that is the flow limit of the edges}

\item{edge_name}{A character string. This is the edge attribute that contains the edge_name of the edges.}

\item{tstep}{A numeric value. The time step, measured in seconds, that will be used to calculate the new dynamic state}

\item{maxIter}{An integer. The maximum number of iterations before stopping. Larger networks usually need more iterations.}

\item{frctmultiplier}{A numeric value. Used to set a multiplier on the friction value. Generally leave this alone..s.}

\item{tol}{A numeric. The tolerance factor for early stopping.}

\item{verbose}{Logical. This value sets whether messages generated during the process are supressed or not.}

\item{friction_stop}{Logical. Includes friction as a stopping condition. useful in some situations}
}
\description{
This function uses the SETS embedding to find the equilibrium position of a network or a bi-connected component
}
