% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindStabilSystem_expanded.R
\name{FindStabilSystem_expanded}
\alias{FindStabilSystem_expanded}
\title{Find stabil system expanded}
\usage{
FindStabilSystem_expanded(NodeStatus, EdgeNode, kvect, dvect, tstep,
  maxIter = 1000, frctmultiplier = 1)
}
\arguments{
\item{NodeStatus}{A data frame The current dynamics and forces experienced by the node a data frame.}

\item{EdgeNode}{}

\item{kvect}{A numeric vector of the spring stiffnesses}

\item{dvect}{A numeric vector of the initial distances between the nodes}

\item{tstep}{A numeric value. The time step, measured in seconds, that will be used to calculate the new dynamic state}

\item{maxIter}{An integer. The maximum number of iterations before stopping. Larger networks usually need more iterations.}

\item{frctmultiplier}{A numeric value. Used to set a multiplier on the friction value. Generally leave this alone..s.}
}
\description{
A helper function to Find_network_balance_expanded()
}
\seealso{
[Find_network_balance_expanded()]
}
