% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_elevation_to_origin.R
\name{fix_elevation_to_origin}
\alias{fix_elevation_to_origin}
\title{fix elevation to origin}
\usage{
fix_elevation_to_origin(relative_blocks, ArticulationVect)
}
\arguments{
\item{relative_blocks}{a data frame containing the outputs from the}

\item{ArticulationVect}{The articulation nodes of the network.}
}
\description{
Internal function. This function makes the elevation of the different blocks all relative to the same fixed point, 
this allows easy use of bi-connected components This is a helper function only.
}
\details{
When a network is broken into bi-connected components the elevation of the nodes is relative to the
local origin. This means all bi-connected components have there own "center of gravity". This function makes sure
all nodes are relative to a fixed point
}
