% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForceV_from_angle.R
\name{ForceV_from_angle}
\alias{ForceV_from_angle}
\title{Calculate for from a given angle}
\usage{
ForceV_from_angle(target_angle = 5 * pi/12, k, d = temp$d)
}
\arguments{
\item{target_angle}{A numeric. The angle for which the force is needed}

\item{k}{A numeric. The spring constant}

\item{d}{A numeric. The original distance between nodes}
}
\description{
This is mostly useful just for single edges. It may be removed from the package for being to specialised
}
