% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise_and_aggregate.R
\name{harmonise_and_aggregate}
\alias{harmonise_and_aggregate}
\title{Harmonise and aggregate}
\usage{
harmonise_and_aggregate(
  vect,
  component_adjust_mat,
  harmonise_vector = TRUE,
  colsum = TRUE
)
}
\arguments{
\item{vect}{a numeric vector, this is the vector that requires harmonising and aggregating}

\item{component_adjust_mat}{a numeric matrix. the output of the adjust_components function}

\item{harmonise_vector}{logical. should the vector be harmonised or just aggregated}

\item{colsum}{a logical. Either the function does a col sum or a col means.}
}
\description{
A helper function of create_stabilised_blocks_expanded poss also create_stabilised_blocks
This is used to quickly aggregate the biconnected component values removing double nodes
It is much faster than using the dplyr summarise function.
It works as the position of all the nodes are already known.
The function outputs a vector
}
