% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_graph.R
\name{hclust_graph}
\alias{hclust_graph}
\title{Hierarchially cluster based on metric.}
\usage{
hclust_graph(g, weight = "weight", method = "single")
}
\arguments{
\item{g}{An igraph object. The graph to analyse}

\item{weight}{A character string. The graph edge attribute that will be clustered}

\item{method}{A character string. The clustering method}
}
\description{
Takes a graph and returns a hierarchical cluster object.
}
