% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_interaction_matrix.R
\name{make_interaction_matrix}
\alias{make_interaction_matrix}
\title{Make interaction network}
\usage{
make_interaction_matrix(target_df, node_component_df)
}
\description{
This is a helper function for the recombine function used by stabilised block, It creates a sparse adjacency matrix
that has the interaction between nodes and the floor/ceiling nodes of the articulation nodes on the shortest path
between the origin node in the origin component and all the other nodes in the network.
}
\details{
The networks created are bipartite

Using the relative blocks is probably not optimal this can be changed when the process works as expected
}
