% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_dc_load.R
\name{normalise_dc_load}
\alias{normalise_dc_load}
\title{Normalise load}
\usage{
normalise_dc_load(g, generation = "generation", demand = "demand",
  net_generation = "net_generation", capacity = "edge_limit",
  edge_name = "edge_name", node_name = "node_name",
  power_flow = "power_flow")
}
\arguments{
\item{g}{An igraph object. Thr graph representation of the network to be normalised}

\item{generation}{A character string. The node attribute containing the flow generation}

\item{demand}{A character string. The node attribute containing the flow demand}

\item{net_generation}{A character string. The node atribute containing the difference between demand and generation for 
the current parametrization}

\item{capacity}{A character string. The edge attribute containing the flow capacity of each edge}

\item{edge_name}{A character string. The edge attribbute containing the edge names}

\item{node_name}{A character string. The node attibute containing the node names}

\item{power_flow}{A character string. The name of the edge atribute that contains the flow data}
}
\description{
This function normalises the load across a DC power network.
}
