% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_data_for_find_network_balance.R
\name{Prepare_data_for_find_network_balance}
\alias{Prepare_data_for_find_network_balance}
\title{Preapre data for the `find network balance` function}
\usage{
Prepare_data_for_find_network_balance(g, force, flow, distance, mass,
  edge_name = edge_name)
}
\arguments{
\item{g}{An igraph object of the network that is going to be turned into a spring embedding.}

\item{force}{a character string. The node attribute that contains the force information for the network.}

\item{distance}{a character string. The edge attribute that contains the distance of the edge.}

\item{mass}{a character string. the node attribute that contains the mass of the nodes.}

\item{edge_name}{a character string. The edge attribute that contains the names of all the edges.}
}
\description{
This is a helpfer function that makes `find_network_balance` code easier to read. It prepares the data into three datasets.
Whether stuff is a character string or not needs to be double checked!
}
\details{
The file outputs a named list containing
}
