% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_articulation_duplicates.R
\name{remove_articulation_duplicates}
\alias{remove_articulation_duplicates}
\title{remove articulation duplicates}
\usage{
remove_articulation_duplicates(node_embeddings, ArticulationVect)
}
\arguments{
\item{node_embeddings}{A dataframe, produced after running fix_z_to_origin}

\item{ArticulationVect}{Vector of articulation nodes}
}
\description{
This is a helper function that is used to efficiently aggregate the articulation nodes after embedding
}
\details{
This function uses vectors to be as fast as possible and use as little memory as possible
}
