% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_small_components.R
\name{remove_small_components}
\alias{remove_small_components}
\title{Remove small components}
\usage{
remove_small_components(g)
}
\arguments{
\item{g}{An igraph object of the graph to embed.}
}
\value{
An igrapg object.
}
\description{
keep only maximum component of graph
}
\details{
As SETSe only works on connected components this function removes all but the largest component. 
This is a helper function to quickly project a network with SETSe.
}
