% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SETS_embedding.R
\name{SETS_embedding}
\alias{SETS_embedding}
\title{calculate SETS embedding}
\usage{
SETS_embedding(g, force = "force", flow = "flow",
  distance = "distance", capacity = "capacity",
  edge_name = "edge_name", k = "k", tstep, coef_drag = 1, tol,
  max_iter = 20000, mass = 1, verbose = FALSE, sparse = FALSE)
}
\arguments{
\item{g}{An igraph object}

\item{force}{A character string. This is the node attribute that contains the force the nodes exert on the network.}

\item{flow}{A character string. This is the edge attribute that is the power flow on the edges.}

\item{distance}{A character string. The edge attribute that contains the original/horizontal distance between nodes.}

\item{capacity}{A character string. This is the edge attribute that is the flow limit of the edges.}

\item{edge_name}{A character string. This is the edge attribute that contains the edge_name of the edges.}

\item{k}{A character string. This is k for the moment don't change it.}

\item{tstep}{A numeric. The time interval used to iterate through the network dynamics.}

\item{coef_drag}{A numeric. This sets the multiplier of friction. Only use if you want to be annoyed and confused}

\item{tol}{A numeric. The tolerance factor for early stopping.}

\item{max_iter}{An integer. The maximum number of iterations before stopping. Larger networks usually need more iterations.}

\item{mass}{A numeric. This is the mass constant of the nodes in normalised networks this is set to 1.}

\item{verbose}{Logical. This value sets whether messages generated during the process are supressed or not.}

\item{sparse}{Logical. Whether sparse matrices will be used. This becomes valubale for larger networks}
}
\value{
A list containing 3 dataframes, the dataframe of the node embeddings, edge embeddings, and network dynamics
}
\description{
A wrapper function that takes a prepared graph and outputs a list of the embeddings and the aggregate dynamics throughout the simulations.
}
