% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SETSe.R
\name{SETSe}
\alias{SETSe}
\title{SETSe algorithm}
\usage{
SETSe(g, force = "net_generation", flow = "power_flow",
  distance = "distance", capacity = "capacity",
  edge_name = "edge_name", k = "k", tstep = 0.02, mass = 1,
  max_iter = 20000, coef_drag = 1, tol = 1e-06, sparse = FALSE,
  two_node_solution = TRUE, include_edges = FALSE, sample = 1)
}
\arguments{
\item{g}{An igraph object}

\item{force}{A character string. This is the node attribute that contains the force the nodes exert on the network.}

\item{flow}{A character string. This is the edge attribute that is the power flow on the edges.}

\item{distance}{A character string. The edge attribute that contains the original/horizontal distance between nodes.}

\item{capacity}{A character string. This is the edge attribute that is the flow limit of the edges.}

\item{edge_name}{A character string. This is the edge attribute that contains the edge_name of the edges.}

\item{k}{A character string. This is k for the moment don't change it.}

\item{tstep}{A numeric. The time interval used to iterate through the network dynamics.}

\item{mass}{A numeric. This is the mass constant of the nodes in normalised networks this is set to 1.}

\item{max_iter}{An integer. The maximum number of iterations before stopping. Larger networks usually need more iterations.}

\item{coef_drag}{A numeric.}

\item{tol}{A numeric. The tolerance factor for early stopping.}

\item{sparse}{Logical. Whether or not the function should be run using sparse matrices. must match the actual matrix, this could prob be automated}

\item{two_node_solution}{Logical. The newton-raphson algo is used to find the correct angle}

\item{include_edges}{Logical. Whether the function will return the edge strain and tension data or just elevation and dynamics. 
Currently default is set to FALSE.}

\item{sample}{Integer. The dynamics will be stored only if the iteration number is a multiple of the sample. 
This can greatly reduce the size of the results file for large numbers of iterations. Must be a multiple of the max_iter}
}
\value{
A list of three elements. A data frame with the height embeddings of the network, a data frame of the edge embeddings
as well as the convergence dynamics dataframe for the network.
}
\description{
This is the basic SETS embeddings algorithm, it outputs all elements of the embeddings as well as convergence dynamics. It is a
wrapper around the core SETS algorithm which requires data preparation and only produces node embeddings and entwork dynamics.
}
\details{
This function is often used in conjunction with \code{Create_stabilised_blocks} and \code{create_balanced_blocks}
}
\seealso{
\code{\link{Create_stabilised_blocks}} \code{\link{create_balanced_blocks}}
}
