% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SETSe_data_prep_orig.R
\name{SETSe_data_prep_orig}
\alias{SETSe_data_prep_orig}
\title{Preapre data for the `SETSe_core` function LEGACY}
\usage{
SETSe_data_prep_orig(
  g,
  force,
  flow,
  distance,
  mass,
  edge_name = edge_name,
  k = "k",
  sparse = FALSE
)
}
\arguments{
\item{g}{An igraph object of the network that is going to be turned into a spring embedding.}

\item{force}{a character string. The node attribute that contains the force information for the network.}

\item{flow}{A character string. This is the edge attribute that is the power flow on the edges.}

\item{distance}{a character string. The edge attribute that contains the distance of the edge.}

\item{mass}{A numeric. This is the mass constant of the nodes in normalised networks this is set to 1.}

\item{edge_name}{a character string. The edge attribute that contains the names of all the edges.}

\item{sparse}{Logical. Whether or not the function should be run using sparse matrices. must match the actual matrix, this could prob be automated}
}
\description{
This is becuase there were issues where an error appears ONLY when the function is  as a package. the problem is summary(Adj)
}
\details{
The file outputs a named list containing
}
