% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_node_solution.R
\name{two_node_solution}
\alias{two_node_solution}
\title{two node solution}
\usage{
two_node_solution(g, Prep = Prep, auto_setse_mode = FALSE)
}
\arguments{
\item{g}{the graph must be two nodes connected by a single edge}

\item{Prep}{The output of SETSe_data_prep. provides the inputs needed to do the two node solution}

\item{auto_setse_mode}{outputs an additional list element "memory_df" to make it compatible with the auto_SETSe function}
}
\value{
A list of two elements node_embeddings and network_dynamics. if `auto_setse_mode==TRUE` 
then a third element is returned the memory_df dataframe
}
\description{
Uses a newton-raphson to solve the two node solution.
}
\details{
a helper function inside the main SETSe suite of functions. However it can be used to solve two node graphs
}
