% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{velocity}
\alias{velocity}
\title{Node velocity}
\usage{
velocity(u, a, t0, t1)
}
\arguments{
\item{u}{A numeric. The velocity at time t0}

\item{a}{A numeric. The acceleration at time t0}

\item{t0}{A numeric. The time at beginning of the time step}

\item{t1}{A numeric. The time at the end of the time step.}
}
\description{
Calculate the velocity of the nodes at time step t. This is a helper function.
}
\examples{
velocity(0, 1, 2,5)
}
