% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_components.R
\name{adjust_components}
\alias{adjust_components}
\title{this is a fiendish internal function that I wish I had described in greater detail}
\usage{
adjust_components(g, max_iter, force)
}
\arguments{
\item{g}{An igraph object of the network}

\item{max_iter}{The maximum number of iterations used in the SETSe process}

\item{force}{A character string  of the variable containing the network force}
}
\description{
This function is only used by Create_stabilised_blocks_expanded. A much more efficient version is used by 
'Create_stabilised_blocks'. If possible this function should be removed and replaced
}
