% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_line_strain.R
\name{Calc_line_strain}
\alias{Calc_line_strain}
\title{Calculate line strain}
\usage{
Calc_line_strain(g, solved_height_df, distance = "distance", capacity,
  flow = "power_flow", edge_name = "edge_name")
}
\arguments{
\item{g}{An igraph object of the network}

\item{distance}{A character string. The name of the edge attribute that contains the distance between two nodes. The default is "distance"}

\item{flow}{A character string. The name of the edge attribute that contains the flow between the two nodes at the end of the edge. The default is "power_flow".}

\item{edge_name}{A character string. The name of the edge attribute that contains the edge name. The default is "edge_name".}

\item{solved_heigh_df}{A data frame. This is the results of Create_stabilised_blocks or Find_network_balance}

\item{capcity}{A character string. The name of the edge attribute that contains the flow capacity of the edge between two nodes.}
}
\description{
This function calculates the line strain characteristics for a graph
}
