% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Tension_matrix.R
\name{Create_Tension_matrix}
\alias{Create_Tension_matrix}
\title{Create tension matrix}
\usage{
Create_Tension_matrix(ten_mat, zvect, zvect_t, dvect, kvect, non_empty_index)
}
\arguments{
\item{ten_mat}{A dataframe. represents the edge node matrix}

\item{zvect}{A numeric vector. The node elevation, this has one element for every edge in the matrix}

\item{zvect_t}{A numeric vector. The node elevation transpose, this has one element for every edge in the matrix}

\item{dvect}{A numeric vector. The original distance between adjacent nodes, this has one element for every edge in the matrix}

\item{kvect}{A numeric vector. The vector of spring coefficients between connected nodes, this has one element for every edge in the matrix}

\item{non_empty_index}{non_empty_matrix A numeric matrix. This contains the node indexes in the adjacency matrix it has 4 columns
the row, the column, the absolute index the transpose of the absolute index.}
}
\description{
Calculates the edge tension in the network. This is a helper function and is not called directly. Ideally the entire thing will be
rewritten in C++
}
