% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_z_to_origin.R
\name{fix_z_to_origin}
\alias{fix_z_to_origin}
\title{fix z to origin}
\usage{
fix_z_to_origin(relative_blocks, ArticulationVect)
}
\arguments{
\item{relative_blocks}{a data frame containing the outputs from the}

\item{ArticulationVect}{The articulation nodes of the network.}
}
\description{
This function makes the elevation of the different blocks all relative to the same fixed point, this allows easy use of bi-connected components
}
\details{
The function needs to be renamed to make it more SETSe friendly something like fix_elevation_to_origin. But I need to see what it will impact first
}
