% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_interaction_matrix.R
\name{make_interaction_matrix}
\alias{make_interaction_matrix}
\title{Make interaction network}
\usage{
make_interaction_matrix(target_df, node_component_df)
}
\arguments{
\item{target_df}{A dataframe... I need to check what of}

\item{node_component_df}{A dataframe representing the embedded nodes of a bi-conneceted component}
}
\description{
This is a helper function and is called by 'adjust_components' which is itself called by 'Create_stabilised_blocks_expanded' 
used by stabilised block, It creates a sparse adjacency matrix that has the interaction between nodes and the 
floor/ceiling nodes of the articulation nodes on the shortest path between the origin node in the origin 
component and all the other nodes in the network.
}
\details{
The networks created are bipartite

Using the relative blocks is probably not optimal this can be changed when the process works as expected
}
