% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_data_for_find_network_balance.R
\name{Prepare_data_for_find_network_balance}
\alias{Prepare_data_for_find_network_balance}
\title{Preapre data for the `find network balance` function}
\usage{
Prepare_data_for_find_network_balance(g, force, flow, distance, mass,
  edge_name = edge_name, sparse = FALSE)
}
\arguments{
\item{g}{An igraph object of the network that is going to be turned into a spring embedding.}

\item{force}{a character string. The node attribute that contains the force information for the network.}

\item{flow}{A character string. This is the edge attribute that is the power flow on the edges.}

\item{distance}{a character string. The edge attribute that contains the distance of the edge.}

\item{mass}{A numeric. This is the mass constant of the nodes in normalised networks this is set to 1.}

\item{edge_name}{a character string. The edge attribute that contains the names of all the edges.}

\item{sparse}{Logical. Whether or not the function should be run using sparse matrices. must match the actual matrix, this could prob be automated}
}
\description{
This is a helpfer function that makes `find_network_balance` code easier to read. It prepares the data into three datasets.
Whether stuff is a character string or not needs to be double checked!
}
\details{
The file outputs a named list containing
}
