% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SETSe_expanded.R
\name{SETSe_expanded}
\alias{SETSe_expanded}
\title{Calculate SETSe expanded version}
\usage{
SETSe_expanded(
  g,
  force = "net_generation",
  distance = "distance",
  edge_name = "edge_name",
  k = "k",
  tstep = 0.02,
  mass = 1,
  max_iter = 20000,
  coef_drag = 1,
  tol = 1e-06,
  sparse = FALSE,
  verbose = TRUE,
  two_node_solution = TRUE
)
}
\arguments{
\item{g}{An igraph object. The network}

\item{force}{A character string}

\item{distance}{A character string. The name of the graph attribute that contains the graph distance}

\item{edge_name}{A character string. This is the edge attribute that contains the edge_name of the edges.}

\item{k}{A character string. This is k for the moment don't change it.}

\item{tstep}{A numeric. The time in seconds that elapses between each iteration}

\item{mass}{A numeric. The mass in kg of the nodes, this is arbitrary and commonly 1 is used.}

\item{max_iter}{An interger. The maximum nuber of iterations before terminating the simulation}

\item{coef_drag}{A numeric. A multplier used to tune the damping. Generally no need to twiddle}

\item{tol}{A numeric. Early termination. If the dynamics of the nodes fall below this value the algortihm will be classed as 
"converged" and the simulation terminates.}

\item{sparse}{Logical. Whether or not the function should be run using sparse matrices. must match the actual matrix, this could prob be automated}

\item{verbose}{Logical value. Whether the function should output messages or run quietly.}

\item{two_node_solution}{Logical. The newton-raphson algo is used to find the correct angle}
}
\description{
This is a special case function which keeps the history of the network dynamics. It is useful for demonstrations. 
or Parametrizing difficult networks
}
